Remove CustomHelperFiles.cpp/h
