+++ clang/lib/DPCT/CMakeLists.txt
@@ -150,7 +150,6 @@ add_clang_library(DPCT
   ExternalReplacement.cpp
   VcxprojParser.cpp
   LibraryAPIMigration.cpp
-  CustomHelperFiles.cpp
   GenMakefile.cpp
   IncrementalMigrationUtility.cpp
   Rules.cpp
