+++ clang/lib/DPCT/ExprAnalysis.cpp
@@ -731,9 +731,10 @@ void ExprAnalysis::analyzeExpr(const MemberExpr *ME) {
         // Similar code in ASTTraversal.cpp
         TmplArg = "<int *>";
       }
-      addReplacement(ME->getMemberLoc(), "get_" + ReplacementStr + TmplArg + "()");
-      requestFeature(
-          MapNames::PropToGetFeatureMap.at(ME->getMemberNameInfo().getAsString()));
+      addReplacement(ME->getMemberLoc(),
+                     "get_" + ReplacementStr + TmplArg + "()");
+      requestFeature(MapNames::PropToGetFeatureMap.at(
+          ME->getMemberNameInfo().getAsString()));
     }
   } else if (BaseType == "textureReference") {
     std::string FieldName = ME->getMemberDecl()->getName().str();
